+++ PCIT_core/internal/llvm_interface/include/class_impls/native_ptr_decls.hpp
@@ -31,6 +31,7 @@ namespace llvm{
 	class DISubprogram;
 	class DILocation;
 	class DIEnumerator;
+	class DIGlobalVariableExpression;
 
 	
 	class Function;